handle nomination years not aligning with calendar years
for some awards and time periods the period when a novel is eligible for an award does not align with the calendar year.
This is especially true of older awards. For now we handle this by setting the award year for each nominee to the mode
of the years it was nominated.